Change uitpas version to 4.0
The version number is also removed from the OpenAPI file because 1) it was incorrect and 2) if we ever do a new major version, we should probably keep the 4.x version on a separate branch and add a new default branch for 5.0. (Which can then be switched between using the branch menu in Stoplight.) Otherwise guides etc will also need to be versioned by file.